admin/requirements.txt
Flask==0.12.3
config==0.4.2
python-decouple==3.1
sentry-sdk[flask]==0.10.2
